churn_prediction_gcp/requirements.txt
streamlit==1.32.2
pandas==2.2.1
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.46.0
protobuf==4.25.3
numpy==1.26.4